add database test& update front-end test.py
