+++ app/tests/index_test.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <title>Front</title>
+    <title>Front end testing</title>
 </head>
 <body>
     <h1>Sprint 1 Input form</h1>
